Improved texture pack system (See desc)
I fixed a few bugs and added a new texture pack selection window.
Texture packs no longer need all of the textures in them, so if there's
a texture that's not in the texturepack, the game will just fallback on
the default one. Also, there is can now be a preview.png image in the
root of the texture pack zip that will be shown on the new texture pack
selection screen. It can be any size, as long as it's square.