Fixed HUGE performance issue (Info in desc)
In the rebuild function of chunks, the checking if a tile was in view or
not was done in local space, not world space, meaning it would only work
at chunk[0,0]. Now it works in all chunks. Rebuild time is a little
longer, but render time is improved greatly.

I think there are still some issues with the frustum calculation, so I
still have to fix those. But I FINALLY FOUND THE ANSWER TO THE HUGE
PERFORMANCE ISSUES IN GLCRAFT! (Isn't that great?)