Improved performance by a LOT (read desc)
I made the activeChunks into a hashmap with the key being a Vector3i
containing the position of the chunk.

The getChunk method used to search through every single chunk until it
found the one it wanted, and that would take a while.

Now, it just has to use activeChunks.get(pos) to get the chunk it wants.

Doing this improves the time of every getTile, setTile, getMeta, and
setMeta call, as well as other things involving getChunk. A notable
improvement is rebuild time, which is a lot better.