Pathfinding! (See desc)
The pathfinding system currently uses Greedy Best-First Search to
pathfind, I will implement A* in the future.

The test entities currently use it for wandering and chasing, it's a bit
glitchy though.

To test the pathfinding, press 'o' to set the destination and 'p' to
pathfind, and it will draw a path of glass from your current location to
the destination. It will avoid pits, walls, etc.