New ore system, not finished (see desc)
There is a new subclass of tile called TileOre.

The tiles that belong to this are stone, iron ore, coal ore, bluestone
ore, and gold ore.

This class makes it easy to customize what item the ore will drop, how
many, and the minimum tool needed to drop the item.

It also makes it easy to specify how often the tile will generate in the
world, but this is not actually used yet.