Starting on the client side of things (see desc)
I'm planning right now for the server connection process to go like
this:

Client sends PacketPlayerLogin to server.

Server sends PacketPlayerLoginResponse to client.

If PacketPlayerLoginResponse.accept == true, client will proceed.

Now client will periodically send PacketPlayerPos to the server, and the
server will mirror that packet to all of its clients, as it will with
many packets.

I also added a Compressor class, to gzip certain packets before they are
sent, namely PacketSendChunk. This way, instead of 10000+ bytes, these
packets can be as little as 3000 bytes gzipped.

I also plan to completely separate the client and server, so even when
playing singleplayer, you'll actually just be connected to a local
server, kind of like Minecraft.

I still have a lot of stuff to do like Entity handling, but I expect to
do that by using the toNBT method of entities and writing the NBT to a
packet through an NBTOutputStream and then interpreting it on the other
side.