New logging system (see desc)
Instead of System.out.println, use GLogger.log or GLogger.logerr or
GLogger.logf to print. All of the LogSources will put their name before
the log, so it is easier to read logs. If the LogSource is SILENT, then
it will only print to GLCraft.log, not the console. If it is NONE, it
will have no prefix.