Connecting to external server works now (see desc)
...But there is no GUI to support it yet.

I tested it manually, and yes, a client can connect to a server and have
multiple players! Sometimes some things can be a bit glitchy, but it
works for the most part.

I'm going to use UDP Multicasting for LAN server discovery, and add a
manual IP connect option. Almost done, Woohoo!