Way Chunks are sent changed (see desc)
Now Chunks are sent on an on-request basis. When the user logs in successfully, the Client will call requestNextChunk() (which sends a PacketRequestChunk() for the next chunk that hasn't been requested yet).

Then, when the server receives one of these packets, it sends the requested chunk to the player.

Then, when the client receives the PacketSendChunk from the server, it will call requestNextChunk() if there is a chunk left to receive. Rinse and repeat.

This way, the server doesn't send all of the chunks at once and potentially overflow the client's socket's buffer.

So it should work on slower computers now (And on Linux, because for some reason the old method worked on Windows on my computer but not Linux.)