New settings system (see desc)
The default settings.properties is in res/properties/settings.properties. If there isn't one already, this will be copied to [GLCRAFTDIR]/settings.properties. If there are new settings added in a new update, then these settings will be automatically added to the new settings.

In the options menu, you can now set a username. It defaults to Player<Random # Between 1 and 10000>.

The only other setting in use right now is the maxFPS setting.