new PacketMultiPacket
This packet contains an array of packets. Ideally they are similar
packets, so when gzipped and sent over, the packet is really small.

Right now I'm using this with sending all the entities over when the
player joins the game, so instead of lots of ~100byte packets, it can be
compressed to around ~200bytes for one packet containing all of the
entity packets.