Actually fixed server crash (see desc)
The crash actually had a chance to happen when playing singleplayer too,
just less so, because the client runs slower than the server because it
has to render stuff too.

The reason it was happening, is that sometimes the server would
sometimes happen to add to the toAdd ArrayList in another thread while
the main thread was iterating over it.

To fix that, instead of using an iterator, I just did

while(!toAdd.isEmpty()){

//add toAdd[0] and then remove it from the arraylist

}

And  the same thing for toRemove.