Scalable GUI (see desc)
Now the GUI is scalable. Now, whenever the display is resized, the
makeElements method in GUIScreen is called. It is also called when the
GUI is shown. The elements are automatically cleared before this method
is called.

Also, Constants.WIDTH and Constants.HEIGHT have been replaced with
Constants.getWidth() and Constants.getHeight().

In order to make a GUIScreen respond properly to resizing, make sure
that the super.update gets called if it overrides the update method, and
make sure to move the initialization of all of the elements into the
makeElements() method. Also, if you have any fields in the class that
involve the width/height, make sure to initialize those as well in the
makeElements method.