Item and Block IDs are now automatic (desc)
Now, you do not need to override the getID method in tiles and items.

The game will automatically assign them IDs in the order they are
initialized, and plugin tiles still use the same id system.

This means that you SHOULD NOT under ANY CIRCUMSTANCES move ANY block or
tile around in the order they are initialized in Tile.class and
Item.class or the game will be screwed up.