Added fancy water waves
Using a fragment shader, the water now has animated waves. They depend on the weather.

They're very subtle during clear weather, but during rain they are a bit faster and deeper.

Also, it's worth mentioning that since the water is using a different shader, it is stored in a different list, so each chunk has three lists to render; solid blocks, translucent blocks, and water.