Infinite World! [UNSTABLE, READ DESC]
DO NOT TRY TO LOAD ANY WORLDS YOU CARE ABOUT WITH THIS VERSION. IT IS VERY UNSTABLE.

Now some talk about what I did.

First off, you must be aware that GLCraft still does not support negative coordinates, so the world will not expand any further negative than the chunk at 0,0.

Next, I'll talk about what I did. So basically, instead of the client requesting all of the chunks at the beginning, it joins and then requests the chunks near it, and the server will send them. There is also a chunk loading/unloading system, so any chunks not near a player will be unloaded (and at the moment, destroyed) and any chunks near a player that are not loaded will be loaded from the save (not implemented yet) or generated. It is also very slow.

Bugs include, but are not limited to:

* Random disconnects from server
* Messages in console about chunks being unloaded that don't exist
* Lag
* Saves getting royally screwed up the butt