Made chunk loading a bit better
I multithreaded lighting, (It's synchronized, obviously) so loading new chunks is a *bit* easier. (Still pretty laggy though.)

Also, GLCraft uses 1GB of RAM at the beginning, and using visualvm, I can see that the used heap space is spiking up to 1GB ish and then going down to around  75MB every few seconds (But the heap, and therefore used memory, stays the same.) and then after a few minutes of the world being loaded, it calms down a bit for some reason. Still gotta figure that out.

It tries to generate trees and stuff, but the logs don't generate for some reason. I'm also not sure how generating trees on chunk borders is going to work. (If the bordered chunk is not loaded)

So yeah, still buggy. Don't use it.